Implement new World class, bigger game World via multiple screens.
The world is now composed of many areas stitched together. You can walk off the edge of one and will be in the neighboring area. Animals will wander freely between areas without issue. Changes, but does not close, #1. Now you will only crash the map if you walk off the edge of an Area that is on the edge of the World--which is sort of easy to do since there's no world map.